monitor_c/requirements.txt
grpcio==1.51.3
protobuf==4.22.0

